Apply input focus selection as part of VDOM patch.
VirtualDOM library provides no guarantee that hooks will be executed after corresponding DOM elements are in the tree. In fact during initial render hooks are executed right after DOM elements are created, but before they are inserted into document. As a workaround we used to delay `selection` and `focus` changes by a tick. But since there is no `Promise` implementation in servo we used shim, that in turn used `setTimeout(cb, 0)`. This caused text fields to act funny on servo if you typed quickly. More specifically `selection` update would be scheduled, but by the time it was executed input value and selection would have changed so selection update would mess things up.

As it turns out there is no need to delay `selection` change at all as selection changes even for nodes that aren’t part of document are applied and kept when elements are inserted into document.

Different workaround has being added to handle `focus`, so that it’s only delayed on first render.